Step one in dynamic background accomplished,
atleast got a radial background and a circle (not really at the centre)
and too big than needed.

Need to get this all paramatrized so the keyboard can be drawn for any
language, any character and as many sectors as needed to represent the
characters. Although, I personally believe more than 5 sectors will be
just fancy to look at and won't be of much use.